feat(money): 플랫폼 통화 USD 통일 (저장·API·UI·Stripe 센트)
- Flyway V37: 금액 컬럼 NUMERIC(14,2), currency 기본 USD, 기존 원화 환산

- 백엔드: BigDecimal, UsdMoney.toStripeCents, PreparePaymentResponse.stripeAmountCents

- 프론트: formatCurrency, USD 소수 입력, 크레딧 표시 분리

Made-with: Cursor

+++ backend/src/main/java/com/audition/platform/api/dto/CreditOrderSummaryResponse.java
@@ -1,5 +1,6 @@
 package com.audition.platform.api.dto;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 
 public class CreditOrderSummaryResponse {
@@ -7,7 +8,7 @@ public class CreditOrderSummaryResponse {
     private String orderNo;
     private String status;
     private String provider;
-    private long amount;
+    private BigDecimal amount;
     private String currency;
     private long credits;
     private long bonusCredits;
@@ -41,11 +42,11 @@ public class CreditOrderSummaryResponse {
         this.provider = provider;
     }
 
-    public long getAmount() {
+    public BigDecimal getAmount() {
         return amount;
     }
 
-    public void setAmount(long amount) {
+    public void setAmount(BigDecimal amount) {
         this.amount = amount;
     }
