+++ backend/src/main/java/com/audition/platform/api/dto/CreditPackageCatalogItemDto.java
@@ -1,5 +1,7 @@
 package com.audition.platform.api.dto;
 
+import java.math.BigDecimal;
+
 /**
  * 스토어(일반 사용자)용 크레딧 패키지 노출. 활성 상품만 반환한다.
  */
@@ -7,7 +9,7 @@ public class CreditPackageCatalogItemDto {
 
     private String id;
     private String name;
-    private long price;
+    private BigDecimal price;
     private long credits;
     private long bonusCredits;
 
@@ -27,11 +29,11 @@ public class CreditPackageCatalogItemDto {
         this.name = name;
     }
 
-    public long getPrice() {
+    public BigDecimal getPrice() {
         return price;
     }
 
-    public void setPrice(long price) {
+    public void setPrice(BigDecimal price) {
         this.price = price;
     }
